+++ server/routes/api/communities.js
@@ -1,6 +1,6 @@
 const router = require('express').Router()
 const mongoose = require('mongoose')
-const utils = require('../../utils/subscriber')
+const utils = require('../../utils/events')
 
 const Community = mongoose.model('Community')
 
@@ -16,13 +16,8 @@ router.get('/:address', async (req, res, next) => {
 })
 
 router.post('/', async (req, res, next) => {
-  // console.log(req.body.community)
   const {receipt} = req.body.community
-  console.log(receipt.events.TokenCreated)
   await utils.processTokenCreatedEvent(receipt.events.TokenCreated)
-  // const community = new Community({...req.body.community, verified: false})
-  // await community.save()
-  // return res.json({data: community})
   return res.json({})
 })
